+++ Blocks.py
@@ -14,8 +14,10 @@ class Blocks:
         for file in files:
             self.blocks[int(list(file)[0])] = Block(file)
         self.resizeTextures((144, 144))
+        self.size = (144, 144)
     
     def resizeTextures(self, size):
+        self.size = size
         for block in self.blocks:
             if block.texture != None:
                 block.resizedTexture = pygame.transform.scale(block.texture, size)